Allow multiple nodes to work on different flames
After an avi file is found to not exist, the script creates an empty file. This will prevent other nodes from attempting the same flame. Unless there is a race condition, but given the ratio of time rendering to checking for file existence, this is sufficiently unlikely. Worst case, some work gets duplicated.