Move playlist generation to a separate script
This allows me to generate the playlist once, which is needed for a multi-node system where different machines are rendering different segments. Since no one machine renders all videos, no machine would create a complete playlist.

Added a target to the Makefile to generate the playlist.

The playlist file is also created in the movies directory now, which I have mapped to a network share.